Cleanup wpgu for PR (#22)
* Initial Cleanup, extract instance creating

* fixup im.rs

* 🧹 Just sweeping up some unclean code 🧹

* 🧹 Just sweeping some stuff under the rug 🧹

* Oh damn i lost my broom

* 🧹 Found it! 🧹

* 📦 Moving stuff around 📦

* Give control flow back to `main` function

* Fix error handling

* Fix aspect ratio

* Remove (not) ppm image generator
