+++ game-engine/Cargo.toml
@@ -6,12 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-raw-window-handle = "0.4"
-wgpu = "0.12.0"
-pollster = "0.2"
-thiserror = "1"
-vek = "0.15.6"
-bytemuck = { version = "1.4", features = ["derive"] }
-tobj = "3.2.1"
-image = "0.24.1"
-log = "0.4.14"
+thiserror = "1.0"
+log = "0.4"
+rendering = { path = "../rendering" }
