+++ game-engine/src/lib.rs
@@ -1,13 +1,5 @@
-pub mod context;
-pub mod error;
-pub mod renderer;
-pub use context::Context;
+mod context;
 
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        let result = 2 + 2;
-        assert_eq!(result, 4);
-    }
-}
+pub use rendering;
+
+pub use context::Context;
