+++ ide/Cargo.toml
@@ -6,11 +6,14 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-env_logger = "0.9.0"
 game-engine = { path = "../game-engine" }
+common = { path = "../common" }
+env_logger = "0.9.0"
 image = "0.24.1"
 log = "0.4.14"
-vek = "0.15.6"
+thiserror = "1.0"
+raw-window-handle = "0.4"
+anyhow = "1.0"
 
 [dependencies.winit]
 version = "0.26"
