+++ rendering/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "rendering"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+common = { path = "../common" }
+raw-window-handle = "0.4"
+wgpu = "0.12"
+pollster = "0.2"
+tobj = "3.2"
+image = "0.24"
+bytemuck = { version = "1.4", features = ["derive"] }
+thiserror = "1.0"
