+++ rendering/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod camera;
+pub mod model;
+pub mod texture;
+
+mod error;
+mod renderer;
+
+pub use camera::Camera;
+pub use error::RenderingError;
+pub use renderer::Renderer;
