+++ rendering/src/model.rs
@@ -1,0 +1,336 @@
+use std::mem;
+use std::ops::Range;
+use std::path::Path;
+use tobj::LoadOptions;
+use wgpu::util::DeviceExt;
+
+use super::texture;
+use crate::error::LoadError;
+use crate::renderer::RawTranslationMatrix;
+
+use common::Transform;
+
+pub type ModelIndex = usize;
+
+pub trait Vertex {
+    fn desc<'a>() -> wgpu::VertexBufferLayout<'a>;
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct ModelVertex {
+    position: [f32; 3],
+    tex_coords: [f32; 2],
+    normal: [f32; 3],
+}
+
+impl Vertex for ModelVertex {
+    fn desc<'a>() -> wgpu::VertexBufferLayout<'a> {
+        wgpu::VertexBufferLayout {
+            array_stride: mem::size_of::<ModelVertex>() as wgpu::BufferAddress,
+            step_mode: wgpu::VertexStepMode::Vertex,
+            attributes: &[
+                wgpu::VertexAttribute {
+                    offset: 0,
+                    shader_location: 0,
+                    format: wgpu::VertexFormat::Float32x3,
+                },
+                wgpu::VertexAttribute {
+                    offset: mem::size_of::<[f32; 3]>() as wgpu::BufferAddress,
+                    shader_location: 1,
+                    format: wgpu::VertexFormat::Float32x2,
+                },
+                wgpu::VertexAttribute {
+                    offset: mem::size_of::<[f32; 5]>() as wgpu::BufferAddress,
+                    shader_location: 2,
+                    format: wgpu::VertexFormat::Float32x3,
+                },
+            ],
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct Material {
+    pub name: String,
+    pub diffuse_texture: texture::Texture,
+    pub bind_group: wgpu::BindGroup,
+}
+
+#[derive(Debug)]
+pub struct Mesh {
+    pub name: String,
+    pub vertex_buffer: wgpu::Buffer,
+    pub index_buffer: wgpu::Buffer,
+    pub num_elements: u32,
+    pub material: usize,
+}
+
+#[derive(Debug)]
+pub struct Model {
+    pub meshes: Vec<Mesh>,
+    pub materials: Vec<Material>,
+}
+
+impl Model {
+    pub fn load<P: AsRef<Path>>(
+        device: &wgpu::Device,
+        queue: &wgpu::Queue,
+        layout: &wgpu::BindGroupLayout,
+        path: P,
+    ) -> Result<Self, LoadError> {
+        let (obj_models, obj_materials) = tobj::load_obj(
+            path.as_ref(),
+            &LoadOptions {
+                triangulate: true,
+                single_index: true,
+                ..Default::default()
+            },
+        )?;
+
+        let obj_materials = obj_materials?;
+
+        // We're assuming that the texture files are stored with the obj file
+        let containing_folder = path.as_ref().parent().ok_or(LoadError::Missing)?;
+
+        let mut materials = Vec::new();
+        for mat in obj_materials {
+            let diffuse_path = mat.diffuse_texture;
+            let diffuse_texture =
+                texture::Texture::load(device, queue, containing_folder.join(diffuse_path))?;
+
+            let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+                layout,
+                entries: &[
+                    wgpu::BindGroupEntry {
+                        binding: 0,
+                        resource: wgpu::BindingResource::TextureView(&diffuse_texture.view),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 1,
+                        resource: wgpu::BindingResource::Sampler(&diffuse_texture.sampler),
+                    },
+                ],
+                label: None,
+            });
+
+            materials.push(Material {
+                name: mat.name,
+                diffuse_texture,
+                bind_group,
+            });
+        }
+
+        let mut meshes = Vec::with_capacity(obj_models.len());
+        for m in obj_models {
+            let mut vertices = Vec::new();
+            for i in 0..m.mesh.positions.len() / 3 {
+                vertices.push(ModelVertex {
+                    position: [
+                        m.mesh.positions[i * 3],
+                        m.mesh.positions[i * 3 + 1],
+                        m.mesh.positions[i * 3 + 2],
+                    ],
+                    tex_coords: [
+                        *m.mesh.texcoords.get(i * 2).unwrap_or(&0.0f32),
+                        *m.mesh.texcoords.get(i * 2 + 1).unwrap_or(&0.0f32),
+                    ],
+                    normal: [
+                        m.mesh.normals[i * 3],
+                        m.mesh.normals[i * 3 + 1],
+                        m.mesh.normals[i * 3 + 2],
+                    ],
+                });
+            }
+
+            let vertex_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                label: Some(&format!("{:?} Vertex Buffer", path.as_ref())),
+                contents: bytemuck::cast_slice(&vertices),
+                usage: wgpu::BufferUsages::VERTEX,
+            });
+            let index_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                label: Some(&format!("{:?} Index Buffer", path.as_ref())),
+                contents: bytemuck::cast_slice(&m.mesh.indices),
+                usage: wgpu::BufferUsages::INDEX,
+            });
+
+            meshes.push(Mesh {
+                name: m.name,
+                vertex_buffer,
+                index_buffer,
+                num_elements: m.mesh.indices.len() as u32,
+                material: m.mesh.material_id.unwrap_or(0),
+            });
+        }
+
+        Ok(Self { meshes, materials })
+    }
+}
+
+pub trait DrawModel<'a> {
+    fn draw_mesh(
+        &mut self,
+        mesh: &'a Mesh,
+        material: &'a Material,
+        camera_bind_group: &'a wgpu::BindGroup,
+    );
+    fn draw_mesh_instanced(
+        &mut self,
+        mesh: &'a Mesh,
+        material: &'a Material,
+        instances: Range<u32>,
+        camera_bind_group: &'a wgpu::BindGroup,
+    );
+
+    fn draw_model(&mut self, model: &'a Model, camera_bind_group: &'a wgpu::BindGroup);
+    fn draw_model_instanced(
+        &mut self,
+        model: &'a Model,
+        instances: Range<u32>,
+        camera_bind_group: &'a wgpu::BindGroup,
+    );
+}
+
+impl<'a, 'b> DrawModel<'b> for wgpu::RenderPass<'a>
+where
+    'b: 'a,
+{
+    fn draw_mesh(
+        &mut self,
+        mesh: &'b Mesh,
+        material: &'b Material,
+        camera_bind_group: &'b wgpu::BindGroup,
+    ) {
+        self.draw_mesh_instanced(mesh, material, 0..1, camera_bind_group);
+    }
+
+    fn draw_mesh_instanced(
+        &mut self,
+        mesh: &'b Mesh,
+        material: &'b Material,
+        instances: Range<u32>,
+        camera_bind_group: &'b wgpu::BindGroup,
+    ) {
+        self.set_vertex_buffer(0, mesh.vertex_buffer.slice(..));
+        self.set_index_buffer(mesh.index_buffer.slice(..), wgpu::IndexFormat::Uint32);
+        self.set_bind_group(0, &material.bind_group, &[]);
+        self.set_bind_group(1, camera_bind_group, &[]);
+        self.draw_indexed(0..mesh.num_elements, 0, instances);
+    }
+
+    fn draw_model(&mut self, model: &'b Model, camera_bind_group: &'b wgpu::BindGroup) {
+        self.draw_model_instanced(model, 0..1, camera_bind_group);
+    }
+
+    fn draw_model_instanced(
+        &mut self,
+        model: &'b Model,
+        instances: Range<u32>,
+        camera_bind_group: &'b wgpu::BindGroup,
+    ) {
+        for mesh in &model.meshes {
+            let material = &model.materials[mesh.material];
+            self.draw_mesh_instanced(mesh, material, instances.clone(), camera_bind_group);
+        }
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct ModelManager {
+    models: Vec<Model>,
+    instances: Vec<Vec<Transform>>,
+    instance_buffers: Vec<wgpu::Buffer>,
+}
+
+impl ModelManager {
+    pub fn new() -> Self {
+        Self {
+            models: vec![],
+            instances: vec![],
+            instance_buffers: vec![],
+        }
+    }
+
+    pub fn add_model(
+        &mut self,
+        device: &wgpu::Device,
+        model: Model,
+        n_instances: u64,
+    ) -> ModelIndex {
+        let idx = self.models.len();
+        self.models.push(model);
+        self.instances.push(vec![]);
+        self.instance_buffers
+            .push(device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some(&format!("Instance buffer {}", self.models.len())),
+                size: n_instances * 4 * 4 * mem::size_of::<f32>() as u64,
+                usage: wgpu::BufferUsages::COPY_DST | wgpu::BufferUsages::VERTEX,
+                mapped_at_creation: false,
+            }));
+        idx
+    }
+
+    pub fn get_transforms(&self, model: ModelIndex, range: Range<usize>) -> &[Transform] {
+        &self.instances[model][range]
+    }
+
+    pub fn modify_transforms_with<F>(
+        &mut self,
+        model: ModelIndex,
+        range: Range<usize>,
+        f: F,
+        queue: &wgpu::Queue,
+    ) where
+        F: FnOnce(&mut [Transform]),
+    {
+        f(&mut self.instances[model][range.clone()]);
+        let raw: Vec<_> = self.instances[model][range.clone()]
+            .iter()
+            .copied()
+            .map(RawTranslationMatrix::new)
+            .collect();
+        queue.write_buffer(
+            &self.instance_buffers[model],
+            range.start as u64 * mem::size_of::<RawTranslationMatrix>() as u64,
+            bytemuck::cast_slice(&raw[..]),
+        );
+    }
+
+    pub fn set_transforms(
+        &mut self,
+        device: &wgpu::Device,
+        queue: &wgpu::Queue,
+        model: ModelIndex,
+        new_transforms: Vec<Transform>,
+    ) {
+        let old_len = self.instances[model].len();
+        let raw: Vec<_> = new_transforms
+            .iter()
+            .copied()
+            .map(RawTranslationMatrix::new)
+            .collect();
+        if old_len < self.instances[model].len() {
+            let new_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                label: Some(&format!("Instance buffer for model {}", model)),
+                contents: bytemuck::cast_slice(&raw),
+                usage: wgpu::BufferUsages::VERTEX,
+            });
+            self.instance_buffers[model] = new_buffer;
+        } else {
+            queue.write_buffer(&self.instance_buffers[model], 0, bytemuck::cast_slice(&raw));
+        }
+        self.instances[model] = new_transforms;
+    }
+
+    pub fn models(&self) -> &[Model] {
+        self.models.as_ref()
+    }
+
+    pub fn instances(&self) -> &[Vec<Transform>] {
+        self.instances.as_ref()
+    }
+
+    pub fn instance_buffers(&self) -> &[wgpu::Buffer] {
+        self.instance_buffers.as_ref()
+    }
+}
